Add button below comment box to mass reply
