+++ src/js/createQuotes.js
@@ -1,0 +1,198 @@
+'use strict';
+
+function loadShitposts() {
+    return [
+        {
+            "name": "contribute",
+            "content": "please contribute copypastas to to https://github.com/HamletDuFromage/4chan-mass-quote/issues"
+        },
+        {
+            "name": "1-Interjection",
+            "content": "I\'d just like to interject for a moment. What you’re referring to as Linux, is in fact, GNU/Linux, or as I’ve recently taken to calling it, GNU plus Linux. Linux is not an operating system unto itself, but rather another free component of a fully functioning GNU system made useful by the GNU corelibs, shell utilities and vital system components comprising a full OS as defined by POSIX.<br><br>Many computer users run a modified version of the GNU system every day, without realizing it. Through a peculiar turn of events, the version of GNU which is widely used today is often called “Linux”, and many of its users are not aware that it is basically the GNU system, developed by the GNU Project. There really is a Linux, and these people are using it, but it is just a part of the system they use.<br><br>Linux is the kernel: the program in the system that allocates the machine’s resources to the other programs that you run. The kernel is an essential part of an operating system, but useless by itself; it can only function in the context of a complete operating system. Linux is normally used in combination with the GNU operating system: the whole system is basically GNU with Linux added, or GNU/Linux. All the so-called “Linux” distributions are really distributions of GNU/Linux."
+        },
+        {
+            "name": "2-No Richard",
+            "content": "No, Richard, it\'s \'Linux\', not \'GNU/Linux\'. The most important contributions that the FSF made to Linux were the creation of the GPL and the GCC compiler. Those are fine and inspired products. GCC is a monumental achievement and has earned you, RMS, and the Free Software Foundation countless kudos and much appreciation.<br><br>Following are some reasons for you to mull over, including some already answered in your FAQ.<br><br>One guy, Linus Torvalds, used GCC to make his operating system (yes, Linux is an OS -- more on this later). He named it \'Linux\' with a little help from his friends. Why doesn\'t he call it GNU/Linux? Because he wrote it, with more help from his friends, not you. You named your stuff, I named my stuff -- including the software I wrote using GCC -- and Linus named his stuff. The proper name is Linux because Linus Torvalds says so. Linus has spoken. Accept his authority. To do otherwise is to become a nag. You don\'t want to be known as a nag, do you?<br><br>(An operating system) != (a distribution). Linux is an operating system. By my definition, an operating system is that software which provides and limits access to hardware resources on a computer. That definition applies wherever you see Linux in use. However, Linux is usually distributed with a collection of utilities and applications to make it easily configurable as a desktop system, a server, a development box, or a graphics workstation, or whatever the user needs. In such a configuration, we have a Linux (based) distribution. Therein lies your strongest argument for the unwieldy title \'GNU/Linux\' (when said bundled software is largely from the FSF). Go bug the distribution makers on that one. Take your beef to Red Hat, Mandrake, and Slackware. At least there you have an argument. Linux alone is an operating system that can be used in various applications without any GNU software whatsoever. Embedded applications come to mind as an obvious example."
+        },
+        {
+            "name": "3-No one:",
+            "content": "no one:<br>absolutely nobody:<br>still nobody:<br>not a single soul:<br>literally no one:<br>not even big chungus:<br>kiwicel: *makes retarded post*<br>leftypol: I\'m going to end this man\'s whole career<br>[everyone disliked that]<br>baby yeed: wait that\'s illegal<br>brie larson: ok that was lowkey on point<br>pickle rick: *slaps roof of car* luke did i ever tell you about the time i turned myself into a pickle? it was an epic moment.<br>luke: *is retarded*<br>CIA: Bane?<br>sans undertale: hey don\'t google HP Lovecraft\'s cat name<br>[OP googles hp lovecrafts cat name]<br>CIA: congratulations you got yourself caught!<br>stan lee: *flies past in a spaceship* ooooh i dont care what universe you\'re from that\'s GOTTA HURT<br>[everyone laughed]<br>keanu reeves: you\'re breathtaking!<br>area 51 guards:i bet i can take on keanu reeves<br>keanu reeves: you sure about that<br>keanu reeves: kills all area 51 guards<br>area 51:wait thats illegal<br>[Everyone liked that]<br>CIA: am I a joke to you?<br>Alt right incels: there\'s no way star wars can be good agai....<br>Baby Yeed: hold my beer<br>Big chungus joined the chat<br>Drumpf has left the chat<br>Markiplier: Welcome To The Dark Side Of Youtube<br>4chanlets: \'Yeah, I\'m thinking this is kind of epic based pilled, maybe a bit of a coom moment?? Idk think I might post a frog."
+        },
+        {
+            "name": "4-India",
+            "content": "Mathematician: How to write 4 in between a 5?<br><br>China: Is this a Joke?<br><br>Japan: Impossible!<br><br>America: The question\'s wrong!!<br><br>UK: Rubbish !!<br><br>India: F(IV)E This is the reason you find Indians everywhere in the world in finance, business, medicine, engineering & arts... anything to do with optimising your brain!!<br><br>British: Can u Swim?<br>Indian: No<br>British: Then a Dog is Better den u because It Swims. Indian: Can u Swim?<br>British: Yes!<br>Indian: Then What\'s the Difference between u & Dog… British Shocked,Faints!! Indian Rocks!<br><br>European : Y do U indians come in all colors, look at us,we R all white..?<br><br>Abdul Kalam: Horses too come in different colors but donkeys R all the same..!!!<br><br>Einstein & a Indian sitting next to each other on a long flight...<br><br>Einstein says:<br>\"Let\'s play a game...<br>I will ask you a question,<br>if you don\'t know the answer,<br>you pay me only $5<br>and<br>if I don\'t know the answer,<br>I will pay you $500...\"<br><br>Einstein asks the first question:<br>What\'s the distance from the Earth to the Moon...?<br><br>Indian doesn\'t say a word,<br>Reaches his pocket,<br>Pulls out a $5...<br><br>Now...<br>It\'s the indian turn...<br><br>He asks Einstein:<br>What goes up a hill with 3 legs<br>and<br>comes down on 4 legs..?<br><br>Einstein searches the net and asks all his smart friends...<br>After an hour he gives Indian $500...<br><br>Einstein going nuts and asks:<br>Well...<br>so what goes up a hill with three legs and comes down with four..?<br><br>Indian reaches his pocket and gives Einstein $5...<br><br>Einstein fainted.....<br><br>Send to all Indians all over the globe! Magic Missing Indian flag in this? Don\'t worry send this message to only 2 groups u will shock to see all these flags will become Indian flags Created by: engineering students Of CIT "
+        },
+        {
+            "name": "5-Navy Seals",
+            "content": "What the fuck did you just fucking say about me, you little bitch? I\'ll have you know I graduated top of my class in the Navy Seals, and I\'ve been involved in numerous secret raids on Al-Quaeda, and I have over 300 confirmed kills. I am trained in gorilla warfare and I\'m the top sniper in the entire US armed forces. You are nothing to me but just another target. I will wipe you the fuck out with precision the likes of which has never been seen before on this Earth, mark my fucking words. You think you can get away with saying that shit to me over the Internet? Think again, fucker. As we speak I am contacting my secret network of spies across the USA and your IP is being traced right now so you better prepare for the storm, maggot. The storm that wipes out the pathetic little thing you call your life. You\'re fucking dead, kid. I can be anywhere, anytime, and I can kill you in over seven hundred ways, and that\'s just with my bare hands. Not only am I extensively trained in unarmed combat, but I have access to the entire arsenal of the United States Marine Corps and I will use it to its full extent to wipe your miserable ass off the face of the continent, you little shit. If only you could have known what unholy retribution your little \"clever\" comment was about to bring down upon you, maybe you would have held your fucking tongue. But you couldn\'t, you didn\'t, and now you\'re paying the price, you goddamn idiot. I will shit fury all over you and you will drown in it. You\'re fucking dead, kiddo."
+        },
+        {
+            "name": "6-Rick & Morty",
+            "content": "To be fair, you have to have a very high IQ to understand Rick and Morty. The humour is extremely subtle, and without a solid grasp of theoretical physics most of the jokes will go over a typical viewer\'s head. There\'s also Rick\'s nihilistic outlook, which is deftly woven into his characterisation- his personal philosophy draws heavily from Narodnaya Volya literature, for instance. The fans understand this stuff; they have the intellectual capacity to truly appreciate the depths of these jokes, to realise that they\'re not just funny- they say something deep about LIFE. As a consequence people who dislike Rick & Morty truly ARE idiots- of course they wouldn\'t appreciate, for instance, the humour in Rick\'s existential catchphrase \"Wubba Lubba Dub Dub,\" which itself is a cryptic reference to Turgenev\'s Russian epic Fathers and Sons. I\'m smirking right now just imagining one of those addlepated simpletons scratching their heads in confusion as Dan Harmon\'s genius wit unfolds itself on their television screens. What fools.. how I pity them. 😂<br><br>And yes, by the way, i DO have a Rick & Morty tattoo. And no, you cannot see it. It\'s for the ladies\' eyes only- and even then they have to demonstrate that they\'re within 5 IQ points of my own (preferably lower) beforehand. Nothin personnel kid 😎"
+        },
+        {
+            "name": "7-Genres",
+            "content": "ABSOLUTE KINOGRAPHY<br>1. Человек с киноаппаратом (1929, Dziga Vertov)<br>2. Суд народов or Sud narodov (1947, Yelizaveta Svilova)<br>3. Новости дня (1954, Dziga Vertov)<br><br>CINEMA<br>1. L\'année dernière à Marienbad (1961, Resnais)<br>2. Voskhozhdenie (1977, Larisa Shepitko)<br>3. The Birth of a Nation (1915, D. W. Griffith)<br><br>FILM<br>1. 黒い河 (1957, Masaki Kobayashi)<br>2. Höstsonaten (1978, Ingmar Bergman)<br>3. Dog Star Man: Part I (1962, Stan Brakhage)<br><br>MOVIES<br>1. Le voyage dans la lune (1902, Georges Méliès)<br>2. Le Dernier Samaritain (1991, Scott)<br>3. Offret (1986, Andrei Tarkovsky)<br><br>FLICKS<br>1. Een hagedis teveel (1960, Verhoeven)<br>2. 七人の侍 (1954, Akira Kurosawa)<br>3. Mauvais Garçons (1995, Bay)<br><br>JOINTS<br>1. Mo\' Better Blues (1990, Lee)<br>2. Joe\'s Bed-Stuy Barbershop: We Cut Heads (1983, Lee)<br>3. Do The Right Thing (1989, Lee)<br>"
+        },
+        {
+            "name": "8-Genres-shitpost",
+            "content": "CINÉMA:<br>1: 夢 (1990, Kurosawa)<br>2: This Night (1965, Fassbinder)<br>3: La morte d\'Isotta (1968, Schröter)<br>4: Reassemblage (1982, Minh Hà)<br>5: Кpaдeцът нa пpacкoви (1964, Radev)<br>6: โรงแรมนรก (1957, Pestonji)<br>7: قمران وزيتونة (2001, Abdelhamid)<br>8: 九月 (1984, Zhuangzhuang)<br>9: Ocъдeни Дyши (1975, Radev)<br>10: Le Fils du Masque (2005, Guterman)<br><br>FILMS:<br>1. Пиcьмa мёpтвoгo чeлoвeкa (1986, Lopushansky)<br>2. Un condamné à mort s\'est échappé (1956, Bresson)<br>3. Guns of The Trees (1961, Mekas)<br>4. Love Streams (1984, Cassavetes)<br>5. 牯嶺街少年殺人事件 (1991, Yang)<br>6. Duelle (1976, Rivette)<br>7. Alice in den Städten (1974, Wenders)<br>8. Лигa cпpaвeдливocти Зaкa Cнaйдepa (2021, Snyder)<br>9. Սայաթ-Նովա (1968, Parajanov)<br>10. Mauvais Sang (1986, Carax)<br><br>MOVIES:<br>1. Barry Lyndon(1975, Kubrick)<br>2. 天国と地獄 (1963, Kurosawa)<br>3. Blue (1993, Jarman)<br>4. Fanny och Alexander (1982, Bergman)<br>5. The Long Day Closes (1992, Davies)<br>6. 天国と地獄 (1963, Kurosawa)<br>7. L\'ascension du chevalier noir (2012, Nolan)<br>8. Providence (1977, Resnais)<br>9. แสงศตวรรษ (2006, Weerasethakul)<br>10. Punishment Park (1971, Watkins)<br><br>OVERALL:<br>Sommaren med Monika (1953, Bergman)<br>牯嶺街少年殺人事件 (1991, 楊)<br>চারুলতা (1964, রায়)<br>Ordet (1955, Dreyer)<br>Les Moissons du Ciell (1978, Malick)<br>Oктябpь «Дecять днeй, кoтopыe пoтpяcли миp (1927, Эйзeнштeйн)<br>红高粱 (1987, 张艺谋)<br>Fear and Desire (1953, Kubrick)<br>Зepкaлo (1975, Tapкóвcкий)<br>01010101 01101110 00100000 0100001101101000 01101001 01100101 0110111000100000 01000001 01101110 0110010001100001 01101100 01101111 01110101(1928, Luis Buñuel)<br>浪華悲歌 (1936, 溝口)<br>Le Révélateur (2002, Philippe Garrel)<br><br>ĶÏŅO:<br>Batman v Superman: Dawn of Justice (2016, Snyder)"
+        },
+        {
+            "name": "9-Top 10",
+            "content": "1. Пиcьмa мёpтвoгo чeлoвeкa (1986, Lopushansky)<br>2. Le Fils du Masque (2005, Guterman)<br>3. Guns of The Trees (1961, Mekas)<br>4. L\'ascension du chevalier noir (2012, Nolan)<br>5. 牯嶺街少年殺人事件 (1991, Yang)<br>6. Duelle (1976, Rivette)<br>7. Pauvre Pierrot (1892, Reynaud)<br>8. Batman v Superman: Dawn of Justice (2016, Snyder)<br>9. Սայաթ-Նովա (1968, Parajanov)<br>10. Mauvais Sang (1986, Carax)"
+        },
+        {
+            "name": "10->mfw thinking of you hurting",
+            "content": "Listen here little baby. You\'re gonna get a lot of hurtful and degrading comments, but that ain\'t what I\'m about. Let me just say, you are perfect the way you are. You hear me sugar? PERFECT. Don\'t ever change. You deserve anything and everything you want. Stay safe for me, baby girl. <br>>mfw thinking of you hurting"
+        },
+        {
+            "name": "11-AWOOOGA",
+            "content": "Hummina hummina hummina bazooooooooing! *eyes pop out* AROOOOOOOOGA! *jaw drops tongue rolls out* WOOF WOOF WOOF WOOF WOOF WOOF WOOF WOOF WOOF WOOF WOOF WOOF WOOF WOOF WOOF *tongue bursts out of the outh uncontrollably leaking face and everything in reach* WURBLWUBRLBWURblrwurblwurlbrwubrlwburlwbruwrlblwublr *tiny cupid shoots an arrow through heart* Ahhhhhhhhhhh me lady... *heart in the shape of a heart starts beating so hard you can see it through shirt* ba-bum ba-bum ba-bum ba-bum ba-bum *milk truck crashes into a bakery store in the background spiling white liquid and dough on the streets* BABY WANTS TO FUCK *inhales from the gas tank* honka honka honka honka *masturabtes furiously* ohhhh my gooooodd~"
+        },
+        {
+            "name": "12-For free",
+            "content": ">he\'s a janitor<br>>on the internet<br>>on an anime imageboard<br>>he does it for free<br>>he takes his \"job\" very seriously<br>>he does it because it is the only amount of power & control he will ever have in his pathetic life<br>>he deletes threads he doesn\'t like because whenever he gets upset he has an asthma attack<br>>he deletes threads he doesn\'t like because they interfere with the large backlog of little girl chinese cartoons he still has to watch<br>>he will never have a real job<br>>he will never move out of his parent\'s house<br>>he will never be at a healthy weight<br>>he will never know how to cook anything besides a hot pocket<br>>he will never have a girlfriend<br>>he will never have any friends"
+        },
+        {
+            "name": "13-MacToddlers BTFO",
+            "content": "MACTODDLERS<br> <br> CONSTANTLY<br> <br> CONTINUALLY<br> <br> UNCEASINGLY<br> <br> UNREPEALABLY<br> <br> UNREMITTINGLY<br> <br> PERPETUALLY<br> <br> PERSISTENTLY<br> <br> PERMANENTLY<br> <br> INVARIABLY<br> <br> INCESSANTLY<br> <br> IRREVOCABLY<br> <br> INTERMINABLY<br> <br> INDELIBLY<br> <br> INDEFINITELY<br> <br> EVERLASTINGLY<br> <br> ENDLESSLY<br> <br> ETERNALLY<br> <br> BLOWN<br> <br> THE<br> <br> FUCK<br> <br> OUT<br> <br> UNTIL THE END OF TIME"
+        },
+        {
+            "name": "14-Soy tropes",
+            "content": ">biblically accurate angels<br>>mature anime<br>>lost media<br>>unconventional christmas movies<br>>uncanny valley<br>>liminal spaces<br>>strangely familiar places<br>>junji ito<br>>akira<br>>neon evangilion<br>>serial experiment lain<br>>so bad its good<br>>the pope exists in the car universe<br>>prequel memes<br>>gaslighting<br>>/mu/core<br>>trust the science<br>>just let people enjoy things<br>>sus<br>>techwear<br>>creed aventus<br>>based<br>>who would win walter white or tony soprano<br>>pirating indie games is immoral<br>>its fun with friends<br>>its a good X game but it falls flat as a Y game<br>>japanese movie posters<br>>this scene of Vito from The Sopranos kind of looks like Star Wars!<br>>/tv/ is a bad board<br>>copium<br>>racism outside /b/ not allowed<br>>christianity/mormonism/scientology is just like a cult<br>>lofi hip hop<br>>plastic love<br>>tomboys<br>>john lennon was a wife beater<br>>nobody:<br>>who hurt you<br>>I don't really care about politics, it makes me unhappy<br>>misinformation spreads quicker than real information<br>>Penguinz0<br>>frankenstein is the scientist not the monster<br>>y'all<br>>reporting soijak spammers isnt reddit<br>>Ed Edd n Eddy took place in purgatory<br>>gaslighting<br>>gatekeeping<br>>food analogies are bad<br>>it’s a wyvern not a dragon<br>>snowpiercer is willy wonka!<br>>thrawn<br>>minimalist movie posters<br>>slowburn psychological A24 horror<br>>lovecraftian<br>>gameplay loop<br>>Burning (2018)<br>>whoah this is just like a real life black mirror<br>>serial killer/true crime documentaries and podcasts<br>>video essay<br>>peer reviewed<br>>bechdel test<br>>movie is actually about capitalism/consumerism/colonialism/police brutality"
+        }
+    ];
+}
+
+function createQuotesString(strArray, format, bottom, characterLimit, maxLines) {
+    let res = "";
+    const strLength = strArray[0].length;
+    let quotesNumber = Math.min(strArray.length, Math.floor(characterLimit / (strLength + 1))); // +1 to account for spaces and linebreaks
+    const offset = bottom * (strArray.length - quotesNumber);
+    if (format) {
+        const cols = Math.floor(quotesNumber / maxLines);
+        if (cols < 1) {
+            for (let i = 0; i < quotesNumber; i++) {
+                res += strArray[i] + "<br>";
+            }
+        }
+        else {
+            let r = Math.floor((quotesNumber - maxLines) / cols) + 1;
+            for (let i = 0 + offset; i < maxLines - r + offset; i++) {
+                res += strArray[i] + "<br>";
+            }
+            for (let i = maxLines - r + offset; i < quotesNumber + offset; i++) {
+                res += strArray[i] + (((i - maxLines - r + offset) % (cols + 1) === 0 || i === quotesNumber + offset - 1) ? "<br>" : " ");
+            }
+        }
+    }
+    else {
+        for (let i = 0 + offset; i < quotesNumber + offset; i++) {
+            res += strArray[i] + " ";
+        }
+        res += "<br>";
+    }
+    return res;
+}
+
+export function getBoardLimits(board) {
+    let maxLines = 100;
+    let characterLimit = 2000;
+    switch (board[1]) {
+        case "pol":
+            maxLines = 70;
+            break;
+        case "v":
+            maxLines = 25;
+            break;
+        case "bant":
+            maxLines = 49;
+            break;
+
+        case "jp":
+            characterLimit = 5000;
+            break;
+        case "lit":
+            characterLimit = 3000;
+            break;
+        case "vt":
+            characterLimit = 5000;
+            break;
+        case "qst":
+            characterLimit = 3000;
+            break;
+        case "vt":
+            characterLimit = 5000;
+            break;
+        case "mlp":
+            characterLimit = 3000;
+            break;
+    }
+    return {
+      maxLines,
+      characterLimit,
+    };
+}
+
+export default function createQuotes(action, selected, format, bttm) {
+    let str = "";
+    let board = window.location.href.match(/boards.4chan(?:nel)?.org\/([a-z]+)\/.*thread.*/);
+    let fourchanx = document.querySelector('html[class~="fourchan-x"') === null ? false : true;
+
+    if (action === "shitpost") {
+        str += loadShitposts()[selected].content;
+    }
+
+    else if (board) {
+        const limits = getBoardLimits(board[1]);
+        const maxLines = limits.maxLines;
+        const characterLimit = limits.characterLimit;
+
+        const posts = document.querySelectorAll('div[class~="postContainer"]:not([data-clone])');
+
+        if (action === "regular" || action === "sneed") {
+            const ids = Array.from(posts).map(e => {
+                return ">>" + e.id.slice(2);
+            });
+            str += createQuotesString(ids, format, bttm, characterLimit, maxLines);
+
+            if (action === "sneed") {
+                str += "sneed";
+            }
+        }
+        else if (action === "dubs") {
+            let dubs = [];
+            for (let i = 0; i < posts.length; i++) {
+                if (/(\d)\1\b/.test(posts[i].id)) {
+                    dubs.push(">>" + posts[i].id.slice(2));
+                }
+            }
+            str += createQuotesString(dubs, format, bttm, characterLimit, maxLines);
+        }
+        else if (action === "kym") {
+            let kym = [];
+            let filename = null;
+            let filenameDOM = null;
+            for (let i = 0; i < posts.length; i++) {
+                if (fourchanx) {
+                    filenameDOM = posts[i].querySelector('div[class~="fileText"] > span[class~="file-info"] > a[target]');
+                }
+                else {
+                    filenameDOM = posts[i].querySelector('[class~="fileText"] > a');
+                }
+                if (filenameDOM !== null) {
+                    filename = filenameDOM.textContent
+                    if (/^\b\w{3}\./.test(filename)) {
+                        kym.push(">" + filename);
+                    }
+                }
+
+            }
+            str += createQuotesString(kym, format, bttm, characterLimit, maxLines);
+        }
+    }
+    else { str += "This is not a recongized 4chan thread"; }
+
+    return str;
+};
