+++ src/manifest.json
@@ -35,6 +35,9 @@
       "js": [
         "js/browser-polyfill.min.js",
         "js/content.js"
+      ],
+      "css": [
+        "css/content.css"
       ]
     }
   ],
